pyproject.toml
[tool.poetry]
name = "bless-server"
version = "0.6.0"
description = "Server API used for Bless app"
authors = ["Iqbal Syifa Mahmuda <bal.mahmuda@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
packages = [
    { include = "bless_server" },
    { include = "manage.py" },
    { include = "dquran" },
    { include = "duser" },
]
exclude = ["dquran/fixtures"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "~4.1.5"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
xhtml2pdf = "^0.2.9"
django-treebeard = "^4.6.0"
psycopg2 = "^2.9.3"
django-modeltranslation = "^0.18.7"
django-htmx = "^1.13.0"
django-extensions = "^3.2.1"
django-widget-tweaks = "^1.4.12"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
